# multithreading & multiprocessing speed upgrade