## Proposed changes

multithreaded the chunk processing because it seemed more suited for the task than multiprocessing. Also used multiprocessing to lever more cpu for large tasks.

## Types of changes

- [ ] Bugfix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] Documentation Update

## Further comments

Original: samples/human-generated/*: 104.92341854103142 seconds; samples/llm-generated/* 113.02261029102374 seconds
Updated, multithreading: samples/human-generated/*: 5.4312 seconds; samples/llm-generated/* 4.0974 seconds
Updated, multithreading & multiprocessing: samples/human-generated/*: 3.5411 seconds; samples/llm-generated/* 2.6611 seconds